chore: make evaluation use graphql
