+++ botfront/imports/startup/client/routes.jsx
@@ -7,14 +7,9 @@ import DocumentTitle from 'react-document-title';
 import { Meteor } from 'meteor/meteor';
 import { Provider } from 'react-redux';
 
-import { Accounts } from 'meteor/accounts-base';
-import { ApolloClient } from 'apollo-client';
-import { InMemoryCache } from 'apollo-cache-inmemory';
-import { HttpLink } from 'apollo-link-http';
-import { onError } from 'apollo-link-error';
-import { ApolloLink, Observable } from 'apollo-link';
 import { ApolloProvider } from 'react-apollo';
 import { ApolloProvider as ApolloHooksProvider } from '@apollo/react-hooks';
+import apolloClient from './apollo';
 
 import TemplatesContainer from '../../ui/components/templates/templates-list/Templates';
 import TemplateContainer from '../../ui/components/templates/template-upsert/Template';
@@ -37,55 +32,6 @@ import SetupLayout from '../../ui/layouts/setup';
 import Project from '../../ui/layouts/project';
 import Index from '../../ui/components/index';
 import store from '../../ui/store/store';
-
-
-const request = async (operation) => {
-    operation.setContext({
-        headers: {
-            // eslint-disable-next-line no-underscore-dangle
-            authorization: Accounts._storedLoginToken(),
-        },
-    });
-};
-
-const requestLink = new ApolloLink((operation, forward) => new Observable((observer) => {
-    let handle;
-    Promise.resolve(operation)
-        .then(oper => request(oper))
-        .then(() => {
-            handle = forward(operation).subscribe({
-                next: observer.next.bind(observer),
-                error: observer.error.bind(observer),
-                complete: observer.complete.bind(observer),
-            });
-        })
-        .catch(observer.error.bind(observer));
-
-    return () => {
-        if (handle) handle.unsubscribe();
-    };
-}));
-
-
-const errorLink = onError(({ graphQLErrors, networkError }) => {
-    if (graphQLErrors) {
-        graphQLErrors.forEach(({ message, locations, path }) => console.log(
-            `[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`,
-        ));
-    }
-    if (networkError) console.log(`[Network error]: ${networkError}`);
-});
-
-const httpLink = new HttpLink({
-    uri: '/graphql',
-    credentials: 'same-origin',
-});
-
-const client = new ApolloClient({
-    link: ApolloLink.from([errorLink, requestLink, httpLink]),
-    cache: new InMemoryCache(),
-});
-
 
 const authenticateProject = (nextState, replace, callback) => {
     Tracker.autorun(() => {
@@ -126,8 +72,8 @@ const authenticateAdmin = (nextState, replace, callback) => {
 Meteor.startup(() => {
     render(
         <DocumentTitle title='Botfront.'>
-            <ApolloProvider client={client}>
-                <ApolloHooksProvider client={client}>
+            <ApolloProvider client={apolloClient}>
+                <ApolloHooksProvider client={apolloClient}>
                     <Provider store={store}>
                         <Router history={browserHistory}>
                             <Route exact path='/setup' component={SetupLayout}>
