+++ botfront/imports/ui/components/nlu/evaluation/queries.js
@@ -1,0 +1,27 @@
+import gql from 'graphql-tag';
+
+export const getValidatedActivity = gql`
+query getValidatedActivity(
+    $modelId: String!
+) {
+    getActivity(
+        modelId: $modelId,
+        validated: true,
+    ) {
+        _id,
+        text,
+        intent,
+        entities {
+            value,
+            entity,
+            start,
+            end,
+            confidence
+        },
+        confidence,
+        validated,
+        createdAt,
+        updatedAt
+    }
+}
+`;
